requirements/base.txt
python-dateutil>=1.5
requests-oauthlib>=0.7
setuptools
python-dotenv
httpx